% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_software_input.R
\name{check_software_input}
\alias{check_software_input}
\title{Maxquant or spectronaut}
\usage{
check_software_input(file)
}
\value{
maxquant or spectrounaut
}
\description{
Maxquant or spectronaut
}
