% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geneList.R
\name{create_geneList}
\alias{create_geneList}
\title{geneList for enrichemnt.}
\usage{
create_geneList(
  data_results = data_results,
  comparison_samples = NULL,
  organism = "org.Hs.eg.db"
)
}
\arguments{
\item{organism}{}
}
\value{

}
\description{
geneList for enrichemnt.
}
