% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contaminants.R
\name{plot_contaminants}
\alias{plot_contaminants}
\title{Title}
\usage{
plot_contaminants(
  proteoInput,
  intensityType = "LFQ",
  softwareUsed = "Spectronaut",
  interactive = FALSE
)
}
\arguments{
\item{intensityType}{}
}
\value{

}
\description{
Title
}
