% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlationly.R
\name{plot_correlationly}
\alias{plot_correlationly}
\title{Plot correlation interactively}
\usage{
plot_correlationly(dep, interactive = TRUE)
}
\arguments{
\item{intensity_type}{}
}
\value{

}
\description{
Plot correlation interactively
}
