% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaply.R
\name{plot_heatmaply}
\alias{plot_heatmaply}
\title{Title}
\usage{
plot_heatmaply(
  dep,
  intensity_type = "LFQ",
  dendogram = "both",
  k_row = 0,
  k_col = 0
)
}
\arguments{
\item{intensity_type}{}
}
\value{

}
\description{
Title
}
