% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram_imputed.R
\name{plot_histogram_imputed}
\alias{plot_histogram_imputed}
\title{Plot_histogram}
\usage{
plot_histogram_imputed(data_to_be_imputed, combined = FALSE)
}
\arguments{
\item{data_to_be_imputed}{}
}
\value{

}
\description{
Plot_histogram
}
