% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_normalization_interactive.R
\name{plot_normalization_interactive}
\alias{plot_normalization_interactive}
\title{Plot Distribution of the data before an after normalization}
\usage{
plot_normalization_interactive(se, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Plot Distribution of the data before an after normalization
}
