% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_improved.R
\name{plot_pca_improved}
\alias{plot_pca_improved}
\title{Title}
\usage{
plot_pca_improved(dep, PC_x = 1, PC_y = 2, label_name = NULL, n = 500)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Title
}
