% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profilely.R
\name{plot_profilely}
\alias{plot_profilely}
\title{Profile Plot}
\usage{
plot_profilely(
  dep,
  centered = FALSE,
  intensity_type = "LFQ",
  color = "#56B4E9",
  angle_labels = 45,
  selected_genes = NULL,
  color_selected = "red",
  plot = TRUE,
  clear_selection = FALSE,
  prof_genes_de = FALSE,
  user_genes_de = NULL,
  color_genes_de = "#800080"
)
}
\arguments{
\item{dep}{}
}
\value{

}
\description{
Profile Plot
}
