% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatterly.R
\name{plot_scatterly}
\alias{plot_scatterly}
\title{Scatter plot}
\usage{
plot_scatterly(
  dep = NULL,
  x_sample = NULL,
  y_sample = NULL,
  intensity_type = NULL,
  alpha = 0.5,
  show_lm = TRUE,
  color = "#56B4E9",
  show_genes_user = FALSE,
  user_genes_de = NULL,
  color_genes_de = "dc143c"
)
}
\arguments{
\item{color_genes_de}{}
}
\value{

}
\description{
Scatter plot
}
