% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volacno Plot}
\usage{
plot_volcano(
  proteomics_results = NULL,
  sample_comparison = NULL,
  alpha = 0.8,
  foldchange_cutoff = 1.5,
  p_value_cutoff = 0.05,
  color_down = "cyan3",
  color_up = "brown2",
  p_adj = TRUE,
  show_genes_user = FALSE,
  user_genes_de = NULL,
  color_genes_de = "#800080",
  coord_x = NULL,
  coord_y = NULL
)
}
\arguments{
\item{p_value_cutoff}{}
}
\value{

}
\description{
Volacno Plot
}
