% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_limma.R
\name{test_limma}
\alias{test_limma}
\title{Test_limma}
\usage{
test_limma(
  se,
  type = c("control", "all", "manual"),
  control = NULL,
  test = NULL,
  design_formula = formula(~0 + condition),
  paired = FALSE
)
}
\arguments{
\item{paired}{}
}
\value{

}
\description{
Test_limma
}
